// packages/@oxen-office/pptx/src/domain/diagram/types.ts
export type DiagramAnimLvlStr = "ctr" | "lvl" | "none";

export type DiagramAnimOneStr = "branch" | "none" | "one";

export type DiagramArrowheadStyle = "auto" | "arr" | "noArr";

export type DiagramAutoTextRotation = "grav" | "none" | "upr";

export type DiagramAxisType =
  | "ancst"
  | "ancstOrSelf"
  | "ch"
  | "des"
  | "desOrSelf"
  | "follow"
  | "followSib"
  | "none"
  | "par"
  | "preced"
  | "precedSib"
  | "root"
  | "self";

export type DiagramAxisTypes = readonly DiagramAxisType[];

export type DiagramBendPoint = "beg" | "def" | "end";

export type DiagramBooleans = readonly boolean[];
// ...
export type DiagramDirection = "norm" | "rev";

export type DiagramElementType =
  | "all"
  | "asst"
  | "doc"
  | "node"
  | "nonAsst"
  | "nonNorm"
  | "norm"
  | "parTrans"
  | "pres"
  | "sibTrans";

export type DiagramElementTypes = readonly DiagramElementType[];

export type DiagramFallbackDimension = "1D" | "2D";

export type DiagramFlowDirection = "col" | "row";

export type DiagramLinearDirection = "fromB" | "fromL" | "fromR" | "fromT";

export type DiagramVariableType =
  | "animLvl"
  | "animOne"
  | "bulEnabled"
  | "chMax"
  | "chPref"
  | "dir"
  | "hierBranch"
  | "none"
  | "orgChart"
  | "resizeHandles";

export type DiagramFunctionArgument = DiagramVariableType;

export type DiagramFunctionOperator = "equ" | "gt" | "gte" | "lt" | "lte" | "neq";

export type DiagramFunctionType =
  | "cnt"
  | "depth"
  | "maxDepth"
  | "pos"
  | "posEven"
  | "posOdd"
  | "revPos"
  | "var";

export type DiagramHierBranchStyle = "hang" | "init" | "l" | "r" | "std";

export type DiagramResizeHandlesStr = "exact" | "rel";

export type DiagramFunctionValue =
  | DiagramAnimLvlStr
  | DiagramAnimOneStr
  | DiagramDirection
  | DiagramHierBranchStyle
  | DiagramResizeHandlesStr
  | boolean
  | number;

export type DiagramGrowDirection = "bL" | "bR" | "tL" | "tR";

export type DiagramHierarchyAlignment =
  | "bCtrCh"
  | "bCtrDes"
  | "bL"
  | "bR"
  | "lB"
  | "lCtrCh"
  | "lCtrDes"
  | "lT"
  | "rB"
  | "rCtrCh"
  | "rCtrDes"
  | "rT"
  | "tCtrCh"
  | "tCtrDes"
  | "tL"
  | "tR";

export type DiagramHueDir = "ccw" | "cw";

export type DiagramIndex1 = number;

export type DiagramInts = readonly number[];

export type DiagramModelId = string;

export type DiagramNodeCount = number;

export type DiagramUnsignedInts = readonly number[];

export type DiagramVariableList = {
  readonly variables: readonly DiagramVariable[];
};
// ...
export type DiagramLayoutNode = DiagramLayoutContent & {
  readonly name?: string;
  readonly childOrder?: DiagramChildOrderType;
  readonly moveWith?: string;
  readonly styleLabel?: string;
};

export type DiagramLayoutContent = {
  readonly variables?: DiagramVariableList;
  readonly algorithm?: DiagramAlgorithm;
  readonly shape?: DiagramLayoutShape;
  readonly presentationOf?: DiagramPresentationOf;
  readonly constraints?: readonly DiagramConstraint[];
  readonly rules?: readonly DiagramRule[];
  readonly forEach?: readonly DiagramForEach[];
  readonly choose?: readonly DiagramChoose[];
  readonly children?: readonly DiagramLayoutNode[];
};

export type DiagramAlgorithmType =
  | "composite"
  | "conn"
  | "cycle"
  | "hierChild"
  | "hierRoot"
  | "lin"
  | "pyra"
  | "snake"
  | "sp"
  | "tx";

export type DiagramAlgorithm = {
  readonly type?: DiagramAlgorithmType;
  readonly params?: readonly DiagramAlgorithmParam[];
};

export type DiagramAlgorithmParam = {
  readonly type?: DiagramParameterId;
  readonly value?: DiagramParameterValue;
};

export type DiagramOutputShapeType = "conn" | "none";

export type DiagramLayoutShapeType = PresetShapeType | DiagramOutputShapeType;

export type DiagramLayoutShape = {
  readonly type?: DiagramLayoutShapeType;
  readonly blipId?: string;
  readonly adjustments?: readonly DiagramAdjust[];
};
// ...
export type DiagramPresentationOf = {
  readonly axis?: DiagramAxisTypes;
  readonly pointType?: DiagramElementTypes;
  readonly count?: DiagramUnsignedInts;
  readonly hideLastTransition?: DiagramBooleans;
  readonly start?: DiagramInts;
  readonly step?: DiagramInts;
};

export type DiagramConstraint = {
  readonly type?: DiagramConstraintType;
  readonly forRelationship?: DiagramConstraintRelationship;
  readonly forName?: string;
  readonly referenceType?: DiagramConstraintType;
  readonly referenceForRelationship?: DiagramConstraintRelationship;
  readonly referenceForName?: string;
  readonly operator?: DiagramBoolOperator;
  readonly value?: string;
  readonly factor?: string;
  readonly max?: string;
  readonly min?: string;
};

export type DiagramRule = {
  readonly type?: string;
  readonly value?: string;
  readonly factor?: string;
  readonly max?: string;
  readonly min?: string;
};

export type DiagramForEach = {
  readonly name?: string;
  readonly axis?: DiagramAxisTypes;
  readonly pointType?: DiagramElementTypes;
  readonly count?: DiagramUnsignedInts;
  readonly hideLastTransition?: DiagramBooleans;
  readonly start?: DiagramInts;
  readonly step?: DiagramInts;
  readonly content: DiagramLayoutContent;
};

export type DiagramChoose = {
  readonly name?: string;
  readonly if?: DiagramIf;
  readonly else?: DiagramElse;
};

export type DiagramIf = DiagramLayoutContent & {
  readonly name?: string;
  readonly function?: DiagramFunctionType;
  readonly argument?: DiagramFunctionArgument;
  readonly operator?: DiagramFunctionOperator;
  readonly value?: DiagramFunctionValue;
};

export type DiagramElse = DiagramLayoutContent & {
  readonly name?: string;
};